Mingw fixes
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/sidplay-2.0.9@222 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/audio/mmsystem/mmsystem.cpp
@@ -69,7 +69,7 @@ Audio_MMSystem::~Audio_MMSystem()
     close();
 }
 
-float *Audio_MMSystem::open (AudioConfig &cfg, const char *)
+short *Audio_MMSystem::open (AudioConfig &cfg, const char *)
 {
     WAVEFORMATEX  wfm;
 
@@ -161,7 +161,7 @@ Audio_MMSystem_openError:
     return NULL;
 }
 
-float*Audio_MMSystem::write ()
+short*Audio_MMSystem::write ()
 {
     if (!isOpen)
     {
@@ -207,7 +207,7 @@ float*Audio_MMSystem::write ()
 }
 
 // Rev 1.2 (saw) - Changed, see AudioBase.h
-float *Audio_MMSystem::reset (void)
+short *Audio_MMSystem::reset (void)
 {
     if (!isOpen)
         return NULL;
